Container building to not override the entries on new containers
This method does not override an existing entry if the same name exists  in the definitions or in any definitions of its parents. 
This way it is possible to change entries defined by other packages as those are build after the main application container is build. The main application container SHOULD be the first to be created and  therefore set any entry that will override the latest containers build.